create Home page and Profile page
