+++ src/App.tsx
@@ -1,50 +1,17 @@
-import { FormEvent, useState } from 'react'
+import { Route, Routes } from 'react-router-dom'
 import './App.css'
-import Greeting from './components/Greeting'
 import Navbar from './components/Navbar'
-import Post from './components/Post'
-import usePosts from './hooks/usePosts'
+import Home from './pages/Home'
+import Profile from './pages/Profile'
 
 function App() {
-  const { posts, isLoading, isPending, createPost } = usePosts()
-  const [newTitle, setNewTitle] = useState<string>('')
-  const [newBody, setNewBody] = useState<string>('')
-
-  const handleSubmit = async (e: FormEvent) => {
-    e.preventDefault()
-    try {
-      await createPost(newTitle, newBody)
-    } catch (error) {
-      console.error(error)
-    }
-  }
-
-  const name: string = 'Bright'
-  const isLoggedIn: boolean = true
-
-  const handleNewTitleChange = (e: React.ChangeEvent<HTMLInputElement>) => setNewTitle(e.target.value)
-  const handleNewBodyChange = (e: React.ChangeEvent<HTMLInputElement>) => setNewBody(e.target.value)
-
   return (
     <div className="App">
       <Navbar />
-      <Greeting name={name} isLoggedIn={isLoggedIn} />
-      <form onSubmit={handleSubmit}>
-        <label>Title</label>
-        <input type="text" value={newTitle} required onChange={handleNewTitleChange} />
-        <label>Body</label>
-        <input type="text" value={newBody} required onChange={handleNewBodyChange} />
-        <button type="submit" disabled={isPending}>
-          {isPending ? 'Pending...' : 'Submit'}
-        </button>
-      </form>
-      <div className="feed-container">
-        {isLoading && posts
-          ? posts.map((post) => {
-              return <Post key={post.id} post={post} />
-            })
-          : 'Loading...'}
-      </div>
+      <Routes>
+        <Route path="/" element={<Home />} />
+        <Route path="/profile" element={<Profile />} />
+      </Routes>
     </div>
   )
 }
