+++ src/components/Greeting.tsx
@@ -1,4 +1,3 @@
-import { useState } from 'react'
 import classes from './Greeting.module.css'
 
 interface IGreetingProps {
@@ -6,17 +5,12 @@ interface IGreetingProps {
   isLoggedIn: boolean
 }
 const Greeting = ({ name, isLoggedIn }: IGreetingProps) => {
-  const [greetingMsg, setGreetingMsg] = useState<boolean>(false)
-  const handleClick = () => {
-    setGreetingMsg(!greetingMsg)
-  }
   return (
     <>
       <div className={classes.card}>
-        <h3>{greetingMsg ? 'Hello!' : 'Welcome!'}</h3>
+        <h3>Welcome!</h3>
         <p>{isLoggedIn ? name : 'Unknown'}</p>
       </div>
-      <button onClick={handleClick}>Change greeting message</button>
     </>
   )
 }
